# Closes #2901: Fix performance regression due to setting ndim field in sym entry