--- a/src/compat/gt-132/ArkoudaSymEntryCompat.chpl
+++ b/src/compat/gt-132/ArkoudaSymEntryCompat.chpl
@@ -67,7 +67,8 @@ module ArkoudaSymEntryCompat {
     this.a = try! makeDistArray((...args), etype);
     init this;
     this.shape = tupShapeString(this.tupShape);
-    this.ndim = N;
+    // commented out for perf reasons, needs further investigation
+    // this.ndim = N;
   }
 
   /*
@@ -88,6 +89,7 @@ module ArkoudaSymEntryCompat {
     this.max_bits=max_bits;
     init this;
     this.shape = tupShapeString(this.tupShape);
-    this.ndim = D.rank;
+    // commented out for perf reasons, needs further investigation
+    // this.ndim = D.rank;
   }
 }