In #2865, the `ndim` field was being set after `init this`, which was resulting in a performance hit for multi locale runs. Commenting that line out for now to restore the performance, since it is only used in currently experimental features.

Scan benchmark on 2 locales of an XC:
| before      | this PR     |
| ----------- | ----------- |
| 42.86 GiB/s | 68.50 GiB/s |

Closes #2901